Use reserved variants of attributes in headers
The variants of C attributes with underscores are reserved identifiers
and thus less likely to be interfered with by macros.  This makes them
more appropriate in header files.

The deprecated attribute was also tried to encapsulate the semaphore
structure, but GCC doesn't seem to implement it correctly: it emits
warnings even when deprecated members are referred to in deprecated
functions.